Consolidate session manager example into Monty
