--- examples/session_manager/agents.toml
@@ -1,46 +1,0 @@
-# Registry of agents the session manager can launch.
-#
-# This is the SINGLE source of truth for "which agents exist". Whatever process starts
-# the session manager (the FastAPI server in examples/session_manager/app.py) reads this
-# file and passes the parsed registry into the manager as its workflow init arg. The
-# manager then serves it back over the `available_agents` query, so the app server can
-# expose it at /api/agents — without hardcoding the agent list.
-#
-# Each agent is one [[agents]] table:
-#   key           short, stable id used by callers to pick an agent (CLI --agent-type,
-#                 and eventually the web UI's agent picker).
-#   workflow_type the agent's registered @workflow.defn name — the only thing the manager
-#                 needs to launch it (every harness agent shares the AgentConfig input).
-#   task_queue    the queue that agent's worker polls. The manager dispatches the child
-#                 session there; callers never supply (or misconfigure) a queue.
-#   label         human-readable name for UI surfaces (sidebar, pickers).
-#   description   one-paragraph summary of what the agent does (shown to humans + LLMs;
-#                 whitespace is collapsed on load, so multi-line strings are fine).
-#
-# This file only enumerates which agents EXIST — it intentionally does NOT pick a
-# "default" agent. Defaulting is a caller's policy decision (the CLI, the web UI), not
-# something baked into the shared registry, so each caller decides for itself.
-#
-# To expose a new agent: register its @workflow.defn on a worker, then add a [[agents]] entry.
-
-[[agents]]
-key = "monty-chat"
-workflow_type = "MontyChatAgent"
-task_queue = "monty-dynamic-agent"
-label = "Monty (Conversational)"
-description = """
-A conversational travel-booking assistant. Chat in plain text; an LLM converses to
-gather your trip details, then WRITES its own pydantic-monty Python script and runs
-it via a tool to orchestrate durable flight/hotel search + booking activities,
-replying with the results."""
-
-[[agents]]
-key = "monty-chat-subagent"
-workflow_type = "MontyChatSubagentAgent"
-task_queue = "monty-dynamic-agent"
-label = "Monty (Conversational, Subagent)"
-description = """
-Same conversational travel-booking experience as Monty (Conversational), but instead
-of running its Python scripts inline it drives the barebones Monty (Dynamic) agent as
-a SUBAGENT — starting a script-runner instance and sending it each script over the
-harness subagent toolset. The first end-to-end demo of agents-as-subagents."""
