--- examples/session_manager/justfile
@@ -1,65 +1,0 @@
-# Scoped justfile for the session manager (examples/session_manager).
-#
-# The session manager is shared infrastructure, not an agent: it hosts the SessionManagerWorkflow
-# and serves the API plus packaged Svelte UI. Run it standalone from this directory
-# (`just <recipe>`), or import these recipes into an example's own justfile so the whole stack is
-# runnable from that example's directory (see examples/monty/justfile).
-#
-# uv installs the project's `examples` dependency group on demand, so there's no separate
-# install step. The Temporal connection and GEMINI_API_KEY come from a .env.local file in the
-# directory you run from — copy the example's .env.example to .env.local (see its README).
-
-set dotenv-filename := ".env.local"
-set dotenv-load := true
-
-# Repo root (examples/session_manager sits two levels below it); `uv run` from there resolves the
-# project's pyproject — its deps, the temporal_agent_harness package, and the examples.
-root := justfile_directory() / ".." / ".."
-ui := root / "ui"
-
-# The Temporal connection comes from TEMPORAL_CONFIG_FILE / TEMPORAL_PROFILE, which the workers
-# and server read via ClientConfig.load_client_connect_config(). Set them in .env.local
-# (auto-loaded above) — see the example's .env.example. Nothing is passed on the command line.
-
-# List available recipes.
-default:
-    @just --list
-
-# Show the Temporal connection these recipes will use (read from .env.local).
-config:
-    @echo "TEMPORAL_CONFIG_FILE = ${TEMPORAL_CONFIG_FILE:-<unset — set it in .env.local; see .env.example>}"
-    @echo "TEMPORAL_PROFILE     = ${TEMPORAL_PROFILE:-default}"
-
-# Start a local Temporal dev server (Web UI: http://localhost:8233; needs the `temporal` CLI).
-temporal:
-    temporal server start-dev
-
-# Run the session-manager worker (hosts SessionManagerWorkflow; launches agents by name).
-session-manager:
-    cd "{{root}}" && uv run --group examples python -m examples.session_manager.worker
-
-# Serve the FastAPI API + built Svelte UI on http://localhost:8000 (needs `session-manager` + an agent worker up).
-server: app-build
-    cd "{{root}}" && uv run --group examples uvicorn examples.session_manager.app:app --host 0.0.0.0 --port 8000 --reload
-
-# Install the shared Svelte UI dependencies.
-app-install:
-    pnpm --dir "{{ui}}" install
-
-# Type-check the shared Svelte UI and run the local Svelte 5 syntax guard.
-app-check:
-    pnpm --dir "{{ui}}" run check
-    pnpm --dir "{{ui}}" run check:svelte5
-
-# Build the shared Svelte UI into temporal_agent_harness/ui/dist for packaging/server use.
-app-build:
-    pnpm --dir "{{ui}}" run build
-
-# Build, test, and create the wheel/sdist in dist/.
-package: app-build app-check
-    cd "{{root}}" && uv run pytest
-    cd "{{root}}" && uv build
-
-# Run the shared Svelte UI on Vite for hot reloading. /api is proxied to the FastAPI server on :8000.
-ui-dev:
-    pnpm --dir "{{ui}}" run dev
